services/api/pyproject.toml
[project]
name = "apex-api"
version = "0.1.0"
description = "APEX Engineering Copilot API"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.8.2",
    "pint==0.24.3",
    "numpy==1.26.4",
    "scipy==1.11.4",
    "scikit-learn==1.4.1.post1",
    "deap==1.4.1",
    "matplotlib==3.9.0",
    "jinja2==3.1.3",
    "weasyprint==61.2",
    "pandas==2.2.1",
    "structlog==24.1.0",
    "orjson==3.10.7",
    "pydantic-settings==2.4.0",
    "redis==5.0.4",
    "asyncpg==0.29.0",
    "aiohttp==3.9.5",
    "httpx==0.27.0",
    "celery==5.4.0",
    "prometheus-fastapi-instrumentator==6.1.0",
    "prometheus-client==0.20.0",
    "slowapi==0.1.9",
    "opentelemetry-sdk==1.26.0",
    "opentelemetry-exporter-otlp==1.26.0",
    "opentelemetry-instrumentation-fastapi==0.47b0",
    "opentelemetry-instrumentation-logging==0.47b0",
    "SQLAlchemy==2.0.34",
    "alembic==1.13.2",
    "pyyaml==6.0.2",
    "openpyxl==3.1.5",
    "psycopg2-binary==2.9.9",
    "minio==7.2.8",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.9",
    "sentry-sdk[fastapi]==2.19.0",
    "supabase==2.3.4",
    "duo-client==5.3.0",
    "bcrypt==4.2.0",
    "tenacity==8.2.3",
    "passlib[bcrypt]==1.7.4",
    "zxcvbn==4.4.28",
]

[tool.uv]
dev-dependencies = [
    "pytest==8.2.2",
    "pytest-asyncio==0.23.7",
    "pytest-benchmark==4.0.0",
    "pytest-cov==5.0.0",
    "httpx==0.27.0",
    "schemathesis==3.26.0",
    "ruff==0.5.6",
    "mypy==1.11.1",
]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "B", "UP"]
extend-exclude = ["alembic/versions/*.py"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
show_error_codes = true
pretty = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=apex",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests (fast, isolated)",
    "contract: marks tests as contract/API schema tests",
    "database: marks tests requiring database connection",
    "determinism: marks tests validating calculation determinism",
    "benchmark: marks tests as performance benchmarks",
]

[tool.coverage.run]
source = ["src/apex"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if TYPE_CHECKING:",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "@abstractmethod",
]

